ai-prediction-service/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==0.19.0
requests==2.26.0
websockets==10.0
pydantic==1.8.2
redis==3.5.3
pika==1.2.0
jaeger-client==4.7.0
prometheus-client==0.11.0
python-consul==1.1.0
kafka-python==2.0.2
tensorflow==2.19.0
numpy>=1.26.0
pandas==1.3.0
scikit-learn==0.24.2
matplotlib==3.4.3
seaborn==0.11.2
plotly==5.1.0
ta-lib==0.4.24 